Added Basic Motion Detector Code
Added basic code which convolves image frames from video with derivative
kernel. Then sets mask to an active value if that the derivative between
frames at that pixel meets a motion threshold. The mask of active motion
pixels is then laid over the original picture to show a motion mask.